+++ src/components/layout/card/Memory.tsx
@@ -13,7 +13,7 @@ type Props = {
 export const Memory: React.FC<Props> = ({ song, artist, image, memory }) => {
   return (
     <div>
-      <Card withBorder={true} className="m-1 hover:opacity-70 sm:m-0 ">
+      <Card withBorder={true} className="hover:opacity-70 sm:m-0 ">
         <Link
           href={{
             pathname: "/article",
